Restricting play arena to only Logged in user
