+++ app/javascript/scavenger_hunt/components/Alert/Alert.jsx
