+++ app/javascript/scavenger_hunt/components/Alert/index.js
