+++ app/javascript/scavenger_hunt/index.js
@@ -1,1 +1,1 @@
-export { App } from './App.jsx'
+export { App } from "./App";
