+++ app/javascript/scavenger_hunt/scenes/PlayArena/PlayArena.jsx
@@ -1,10 +1,20 @@
 import React from 'react'
+import { useCurrentUserQuery } from '../../data/queries';
+
 import { Progress } from '../../components/Progress'
 import { Navigation } from './Navigation';
 import { MainContent } from './MainContent';
 import './PlayArena.scss'
 
-const PlayArena = () => {
+const PlayArena = (props) => {
+  const { loading: queryLoading } = useCurrentUserQuery();
+
+  if (queryLoading) {
+    return(
+      <h3>Yet Loading...</h3>
+    )
+  }
+
   return (
     <div className="main">        
       <Progress />
